requirements.txt
fastapi==0.85.0
pydantic==1.10.2
pymongo==4.2.0
python-dotenv==0.20.0
typing-extensions==4.4.0